+++ src/shelf.js
@@ -2,7 +2,7 @@ const { checker } = require("@herbsjs/herbs")
 const { entity2diagram, usecase2diagram } = require('@herbsjs/herbs2mermaid')
 const generateHTML = require('./template/default')
 
-const generateShelfData = (usecases, specs = []) => {
+const generateShelfData = (usecases, specs = [], REST = []) => {
 	const shelfData = []
 	const groups = [...new Set(usecases.map((i) => i.tags.group))]
 	for (const group of groups) {
@@ -13,7 +13,8 @@ const generateShelfData = (usecases, specs = []) => {
 				.map((i) =>
 					formatUseCaseDoc(
 						i.usecase.doc(),
-						specs.find((s) => s.usecase === i.id)
+						specs.find((s) => s.usecase === i.id),
+						REST.find((r) => r.id === i.id)?.REST
 					)
 				)
 		})
@@ -21,7 +22,7 @@ const generateShelfData = (usecases, specs = []) => {
 	return shelfData
 }
 
-const formatUseCaseDoc = (usecase, spec) => {
+const formatUseCaseDoc = (usecase, spec, REST) => {
 	const requestParams = []
 	const responseParams = []
 
@@ -39,7 +40,8 @@ const formatUseCaseDoc = (usecase, spec) => {
 	}
 
 	if (!checker.isEmpty(usecase.response)) {
-		if (Object.entries(usecase.response).length == 0) responseParams.push({ name: 'Object of', type: usecase.response.name })
+		if (Object.entries(usecase.response).length == 0)
+			responseParams.push({ name: 'Instance of', type: usecase.response.name })
 		else {
 			Object.entries(usecase.response).map(([key, value]) => {
 				if (Array.isArray(value)) {
@@ -52,21 +54,38 @@ const formatUseCaseDoc = (usecase, spec) => {
 		usecase.response = responseParams
 	}
 
-	if (spec) {
-		usecase.spec = spec.spec.doc()
+	if (spec) usecase.spec = spec.spec.doc()
+	if (REST) {
+		usecase.REST = JSON.parse(JSON.stringify(REST))
+		function stringify(obj) {
+			function convertFieldsArrayToObject(fields) {
+				const result = {}
+				fields.forEach(field => { result[field.name] = field.type })
+				return result
+			}
+			const result = {}
+			for (const key in obj) {
+				if (Array.isArray(obj[key])) result[key] = obj[key].map(stringify)
+				else if (typeof obj[key] === 'object' && obj[key] !== null) result[key] = stringify(obj[key])
+				else result[key] = obj[key].name
+			}
+			return result
+		}
+		REST.map((endpoints) => { usecase.REST.find((e) => e.path === endpoints.path).parameters = stringify(endpoints.parameters) })
 	}
-
 	return usecase
 }
 
-function renderHTML({ project, usecases, entities, specs, description, readmePath }) {
-	const shelfData = generateShelfData(usecases, specs)
+function renderHTML({ project, usecases, entities, specs, REST, description, readmePath }) {
+	const shelfData = generateShelfData(usecases, specs, REST)
 	const classDiagram = entity2diagram(entities)
 	const usecasesFlowChart = usecase2diagram(usecases)
 	return generateHTML(project, shelfData, description, readmePath, classDiagram, usecasesFlowChart)
 }
 
 function renderShelfHTML(project, usecases, entities, description = '', readmePath = './README.md') {
+	// eslint-disable-next-line no-console
+	console.warn(`⚠️  'renderShelfHTML' function is deprecated. Use the 'herbsshelf' function instead.`)
 	const shelfData = generateShelfData(usecases)
 	const classDiagram = entity2diagram(entities)
 	const usecasesFlowChart = usecase2diagram(usecases)
@@ -92,7 +111,12 @@ function herbsshelf({ herbarium, project, description = '', readmePath = './READ
 		usecase: item.usecase
 	}))
 
-	return renderHTML({ project, usecases, entities, specs, description, readmePath })
+	const REST = Array.from(herbarium.usecases.all).map(([_, item]) => ({
+		id: item.id,
+		REST: item.REST
+	}))
+
+	return renderHTML({ project, usecases, entities, specs, REST, description, readmePath })
 }
 
 module.exports = { renderShelfHTML, herbsshelf }