+++ src/template/components/responseCard.js
@@ -1,15 +1,25 @@
-const ResponseCard =`
+const ResponseCard = `
   {selectedPage.response && selectedPage.response.length > 0 ?
     <div className="content-card">      
       <div className="header-card">
-        <h4 className="icon i-response">Response:</h4>
+        <h4>
+          <span className="icon i-response"></span>
+          Response
+        </h4>
       </div>
       <ul className="steps-list">
       {selectedPage.response.map( (resItem, index) => {
-          if (['Array of', 'Object of'].includes(resItem.name))
-            return <li> {resItem.name} <b>{resItem.type}</b></li>
+          if (['Array of', 'Instance of'].includes(resItem.name))
+            return <li> {resItem.name} 
+              <b> 
+                <span className="response-type"> {resItem.type} </span>
+              </b>
+            </li>
           else
-            return <li> <b>{resItem.name}</b> : {resItem.type}</li>
+            return <li> 
+              <b>{resItem.name}</b> : 
+              <span className="response-type"> {resItem.type} </span> 
+            </li>
         }
       )}
       </ul>