+++ src/template/components/scenarioCard.js
@@ -1,31 +1,45 @@
 const scenario = `
   {selectedPage.spec && selectedPage.spec.scenarios && selectedPage.spec.scenarios.length > 0 ?
-    <div class="content-row">
       <div className="content-card">
         <div className="header-card">
-          <h4 className="icon i-scenario">Scenarios:</h4>        
+          <h4>
+            <span className="icon i-scenarios"></span>
+            Scenarios
+          </h4>        
         </div>        
         
         {selectedPage.spec.scenarios.map( (scenario, _) => {
           return <ul className="scenario-list">
-            🟒 {scenario.description}
+            <span className="icon i-scenario"></span>
+            {scenario.description}
             {scenario.samples.map( (sample, _) => {
-              return <li className="scenario-list">⇶ {sample.description} <div className="scenario-type">(sample)</div></li>
+              return <li className="scenario-list"> 
+                <span className="icon i-sample"></span>
+                {sample.description} <div className="scenario-type">(sample)</div>
+              </li>
             })}
             {scenario.givens.map( (given, _) => {
-              return <li className="scenario-list">🠆 {given.description} <div className="scenario-type">(given)</div></li>
+              return <li className="scenario-list">
+                <span className="icon i-given"></span>
+                {given.description} <div className="scenario-type">(given)</div>
+              </li>
             })}
             {scenario.whens.map( (when, _) => {
-              return <li className="scenario-list">🠆 {when.description} <div className="scenario-type">(when)</div></li>
+              return <li className="scenario-list">
+                <span className="icon i-when"></span>
+                {when.description} <div className="scenario-type">(when)</div>
+              </li>
             })}
             {scenario.checks.map( (check, _) => {
-              return <li className="scenario-list">🠆 {check.description} <div className="scenario-type">(check)</div></li>
+              return <li className="scenario-list">
+                <span className="icon i-check"></span>
+                {check.description} <div className="scenario-type">(check)</div>
+              </li>
             })}
         </ul>
         }
       )}
     </div>
-  </div>
   : null}
 `